requirements.txt
lottie==0.7.1
pillow==11.0.0